Make reconstruction infallible
There were two consistency checks during reconstruction: the initial
code was checked against the table length and then the actual decoded
length was checked against the number of code words. However, both are
inner invariants and not exposed to the public interface.

The first is already checked as part of decoding, only code words
smaller than the next code are reconstructed. The next code is also the
current table length.

Assuming a correct implementation, the second follows from the fact that
the code word tree forms a graph without cycles. This is ensured by the
way in which new codes are created. Violating this property would be a
bug in the implementation and it should not return a user facing error.