Hide the decode state behind dyn trait interface
This is supposed to prepare adhering to parameter on construction.
Instead of the old interface, which exposes them as type parameters,
this crate chooses to use the same type and interface and instead to a
dynamic dispatch, which isn't really dynamic since it always calls the
same implementation.

The internals will nevertheless have a type parameter that implements
the different strategies for decoding byte order and switching but this
detail is hidden from the caller.

This change had no discernible performance difference.