Fix full nodes and resetting clearing code
A full node would pretend to always have arbitrary continuation bytes
already but would link to code 0 for this. This would make it emit a
null byte instead of the actual content.

The reset would also not set the current word to the byte the caused it
which would also forget that byte.