Move async blocks to separate file, avoid parsing
This is a fix only for maximum 1.34.2 compatibility. By moving the full
impl block not only into a separately cfg-gated block but also into a
different file, the expansion and parsing of its code never happens.
This avoids tripping over the usage of `.await` which would technically
be a keyword in old versions/editions. This does not affect the
interface in any regard since those files merely contain impl blocks.