Allow arbitrarily low decode size
While okay for correctness (those sizes do not roundtrip with any
implementation) this is not truly necessary to enforce. It should also
not be enforced with a panic as this makes the interface a bit brittle
(see https://github.com/image-rs/image-gif/issues/96).

This should later be address with a newtype around the CodeSize but that
would be an interface break. It's also not entirely clear if we can
fixup the roundtrip behavior for the code size of 0/1 or not, at least
for our own library. And if not how we best enforce different allowed
sizes for encode and decode.