Simplify condition of code loop break
This pessimizes the case of restarting the loop, as last the word needs to
be reconstructed for writing it out and deriving codes, but since this is
the hottest path it actually speeds up decoding most of the time by
reducing instructions taken. Also the optimizer is more kinds to this
condition.

This opens up the question of whether the difference in behavior is worth
it. Of course, making the flag explicit of non-standard and incomplete
checks may be good on its own and the flag mechanism is also neat.