Reduce work for creating new links
When creating a new code, we create 'link' in an index linked list that
contains the last byte of the new code and the index of its previous code.
We had calculated that last byte anew each time but also redundantly
tracked.

Now, we track exactly that required information, the code and its first
byte in the linked list structure. This also eliminates a data dependency
in the main restoration loop.

Yields around ~5-8% performance in many cases.